🔨Refactor : 타입 수정
기존에 Int 타입을 사용하던 변수들을 Long 타입으로 변경

Ref: #8

+++ src/main/kotlin/com/teamsparta/buysell/domain/post/model/Post.kt
@@ -24,7 +24,7 @@ class Post(
     var content: String,
 
     @Column(name = "price")
-    var price: Int,
+    var price: Long,
 
     @Column(name = "is_soldout")
     var isSoldOut: Boolean = false,
